+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v3.15.0
 ## 🚀 Feature
 - `AndesCarousel` supports XSmall margin and custom horizontal padding
+## 🛠 Fixes
+- Fix `AndesCheckbox` overlap of layouts, set ellipsize end and the support of multi lines | Author: [@mingkao-meli](https://github.com/mingkao-meli)
 
 # v3.14.1
 ## 🛠 Fixes